// gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/api-score/types/asyncJob.ts
/*
 * Copyright © 2015 The Gravitee team (http://gravitee.io)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type AsyncJobType = 'FEDERATED_APIS_INGESTION' | 'SCORING_REQUEST';

export type AsyncJobStatus = 'ERROR' | 'PENDING' | 'SUCCESS' | 'TIMEOUT';

export interface AsyncJob {
    id: string;
    sourceId: string;
    environmentId: string;
    initiatorId: string;
    type: AsyncJobType;
    status: AsyncJobStatus;
    errorMessage?: string;
    createdAt: string;
    updatedAt: string;
}

export interface AsyncJobsResponse {
    data: AsyncJob[];
}

export interface ListAsyncJobsQuery {
    type?: AsyncJobType;
    status?: AsyncJobStatus;
    sourceId?: string;
    page?: number;
    perPage?: number;
}